feature 233: confirmation for deletion of remote sites, user groups and virtual categories
git-svn-id: http://piwigo.org/svn/trunk@1045 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ template/yoga/admin/cat_list.tpl
@@ -21,7 +21,7 @@
         <li><a href="{category.permissions.URL}" title="{lang:edit category permissions}" ><img src="{themeconf:icon_dir}/category_permissions.png" class="button" alt="{lang:permissions}" /></a></li>
         <!-- END permissions -->
         <!-- BEGIN delete -->
-        <li><a href="{category.delete.URL}" title="{lang:delete category}"><img src="{themeconf:icon_dir}/category_delete.png" class="button" alt="{lang:delete}" /></a></li>
+        <li><a href="{category.delete.URL}" title="{lang:delete category}" onclick="return confirm('{lang:Are you sure?}');"><img src="{themeconf:icon_dir}/category_delete.png" class="button" alt="{lang:delete}" /></a></li>
         <!-- END delete -->
       </ul>
