+++ template/yoga/admin/group_list.tpl
@@ -21,7 +21,7 @@
     <td><a href="{group.U_MEMBERS}">{group.MEMBERS}</a></td>
     <td style="text-align:center;">
       <a href="{group.U_PERM}"><img src="{themeconf:icon_dir}/permissions.png" class="button" style="border:none" alt="permissions" title="{lang:permissions}" /></a>
-      <a href="{group.U_DELETE}"><img src="{themeconf:icon_dir}/delete.png" class="button" style="border:none" alt="delete" title="{lang:delete}" /></a>
+      <a href="{group.U_DELETE}" onclick="return confirm('{lang:Are you sure?}');"><img src="{themeconf:icon_dir}/delete.png" class="button" style="border:none" alt="delete" title="{lang:delete}" /></a>
     </td>
   </tr>
   <!-- END group -->
